Make JetBrains.Annotations Private (#5151)
* Format Directory.Packages.props for readability

Indentation was adjusted in the Directory.Packages.props file to enhance readability and maintain a consistent style. No functional changes or impact on package versions were made.

* Update JetBrains.Annotations to private assets

The change in the commit modifies the JetBrains.Annotations package reference in Directory.Build.props to be included as private assets. This adjustment ensures that the JetBrains.Annotations won't be exposed publicly, increasing the security and the overall robustness of the system.

* Added missing references

---------

Co-authored-by: Raymond den Haan <raymond.den.haan@nexxbiz.io>