Remove unused BulkExtensions (#5160)
* Remove unnecessary whitespace in packages.yml

The whitespace after the if conditional in packages.yml was unused and unnecessary. This commit ensures to remove those to promote cleaner, more efficient code.

* Update default package version in GitHub workflows

The package version in the GitHub workflows configuration (.github/workflows/packages.yml) has been updated from 3.1.0 to 3.2.0. This change reflects version updates in the package management system.

* Remove unused BulkExtensions

---------

Co-authored-by: Sipke Schoorstra <sipkeschoorstra@outlook.com>